requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
websockets==11.0.3
requests==2.31.0
beautifulsoup4==4.12.2
chardet==5.2.0
importlib-metadata==6.8.0
Werkzeug==2.3.7
PyJWT==2.8.0
Flask-SQLAlchemy==3.1.1
mysqlclient==2.2.1
python-dotenv==1.0.0
webdavclient3==3.15.0 